Summarise c-square data to a lower resolution

This function acts very much like dplyr::summarise, but instead of using a column to group, c-square codes are used to aggregate to a lower resolution and summarise the data to that lower resolution.

summarise.csquares(x, ..., .by, tiers_down = 1L)

Arguments

x

The object to be summarised. Can be a data.frame, sf, or stars object.

...

<data-masking> Name-value pairs of summary functions. The name will be the name of the variable in the result.

The value can be:

A vector of length 1, e.g. min(x), n(), or sum(is.na(y)).

A data frame, to add multiple columns from a single expression.

Returning values with size 0 or >1 was deprecated as of 1.1.0. Please use reframe() for this instead.

.by

The column name that holds the c-squares codes that need to be aggregated.

tiers_down

The number of tiers down from the current resolution to which you wish to summarise. If the current resolution is 5x5 degrees, the tier down would be 10x10 degrees (as is the case in the example below).

Value

Returns the summarised object inheriting its class from x

Author

Pepijn de Vries

Examples

library(dplyr)
orca |>
  as_csquares(csquares = "csquares") |>
  summarise(
    .by = "csquares",
    orcinus_orca = any(na.omit(.data$orcinus_orca)))
#> # A tibble: 560 × 2
#>    csquares   orcinus_orca
#>  * <csquares> <lgl>       
#>  1 1004       FALSE       
#>  2 1005       FALSE       
#>  3 1006       FALSE       
#>  4 1007       FALSE       
#>  5 1008       FALSE       
#>  6 1009       FALSE       
#>  7 1010       FALSE       
#>  8 1011       FALSE       
#>  9 1012       FALSE       
#> 10 1013       FALSE       
#> # ℹ 550 more rows